requirements.txt
Flask>=3.0.0,<4.0.0
Werkzeug>=3.0.0,<4.0.0
pymongo>=4.6.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0
